Restore Gpu frequency 512 Mhz
At the same time, this patch enables the possibility to change
the  'frecuency_limit' value returned by the PVR driver when
Gpu's overclock interface is changed.

For practical purposes this is not necessary, since this value
has no influence on the graphics performance, but it's always nice
that the operating frequency is displayed on a query to this value.