Input: add infrastructure for selecting clockid for event time stamps Since wall time can jump backwards, it is difficult to use for input because one cannot determine if one event occurred before another or for how long a key was pressed. However, the timestamp field is part of the kernel ABI, and cannot be changed without possibly breaking existing users.
This patch adds a new IOCTL that allows a clockid to be set in the evdev_client
struct that will specify which time base to use for event timestamps
(ie: CLOCK_MONOTONIC instead of CLOCK_REALTIME).
For now we only support CLOCK_MONOTONIC and CLOCK_REALTIME, but in the future
we could support other clockids if appropriate.

The default remains CLOCK_REALTIME, so we don't change the ABI.