Vfs: check if f_count is 0 or negative
filp_close is using !file_count(filp) to check if f_count is 0.
If it is 0, filp_close think it is a closed file then will return.
However, for a closed file, f_count could be reduced to -1, then
!file_count(filp) is false, filp_close will proceed to handle this
file then could panic.
This change will check if f_count is 0 or negative instead of only
checking 0 to avoid panic.